Update ProfilePage layout and button texts in MainWindow
In ProfilePage.xaml, added a new grid layout with two rows. The first row contains four columns with Borders and StackPanels displaying labels for test and question counts. The second row includes a Border with a StackPanel containing multiple Grid elements for user information such as name, phone number, and password options.

In MainWindow.xaml, updated the button texts for TestBtn and ProfileBtn from "Testlaringiz" to "Testlar" and from "Profilingiz" to "Profil", respectively.